+++ altair_saver/__init__.py
@@ -1,9 +1,10 @@
 """Tools for saving altair charts"""
-from ._core import render, save
+from ._core import render, save, available_formats
 from .savers import Saver, BasicSaver, HTMLSaver, NodeSaver, SeleniumSaver
 
 __version__ = "0.4.0.dev0"
 __all__ = [
+    "available_formats",
     "render",
     "save",
     "Saver",